pubmed-multi-agent/requirements.txt
google-adk>=1.24.0
google-cloud-aiplatform[agent_engines,a2a]>=1.127
vertexai
google-cloud-bigquery
google-cloud-bigquery-storage
pandas>=2.2.3
db-dtypes
pydantic
uvicorn
